+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Increase read timeout in Proxmox plugin [PR #2545]
+- systemtests check runscript failed - improve media_vault [PR #2568]
 
 ### Documentation
 - docs: add tape speed test, tapestat & sg_logs in troubleshooting [PR #2555]
@@ -2229,4 +2230,5 @@ If you want to migrate from your manually configured disk autochanger to simply 
 [PR #2549]: https://github.com/bareos/bareos/pull/2549
 [PR #2555]: https://github.com/bareos/bareos/pull/2555
 [PR #2560]: https://github.com/bareos/bareos/pull/2560
+[PR #2568]: https://github.com/bareos/bareos/pull/2568
 [unreleased]: https://github.com/bareos/bareos/tree/master